refactor of common functional utils
